Major new functionality added to grammar class.
Grammars can now contain a new keyword of "GE_ADFS:n", where n is an integer. The new additions to the grammar class will then replace this entry with a set of calls to a list of n ADFs, i.e. adfs[0]|adfs[1]|...|adfs[n-1]. Note that this is implemented in the same mannner as the GE_RANGE:n keyword.

The grammar itself then gets re-written to automatically produce a list of n ADFs. The list of ADFs is defined at the start of each phenotype string. By default, ADFs are given full access to the entire original grammar, i.e. ADFs are built from the start rule of the grammar (this will probably be changed in later implementations).

All grammars which use ADFs in this manner are automatically converted into a .pybnf Python grammar (if they are not already). This requires the fitness function to handle python programs (if it doesn't do so already). To help out with this, a new "eval_or_exec" function has been added to the utilities. This can be called by the fitness function to automatically either evaluate or execute a phenotype string as appropriate. An example of this whole system in use can be seen with the new "adfs" grammars, parameters files, and fitness functions.

Note that the new "eval_or_exec()" function is messy, as it requires the input and subsequent setting of local variables for the desired evaluation/execution environment. Hopefully we can clean this up somewhat in later iterations.